# Use source-map@^0.5.0 and optimize bundle builds