--- a/package.json
+++ b/package.json
@@ -3,8 +3,10 @@
   "devDependencies": {
     "babel": "5.8.19",
     "babel-eslint": "^4.0.6",
-    "browserify": "^11.0.0",
+    "browserify": "^11.2.0",
+    "bundle-collapser": "^1.2.1",
     "chai": "^2.2.0",
+    "derequire": "^2.0.2",
     "es5-shim": "^4.1.7",
     "eslint": "^1.1.0",
     "fs-readdir-recursive": "^0.1.2",