--- a/packages/babel-cli/package.json
+++ b/packages/babel-cli/package.json
@@ -18,7 +18,7 @@
     "output-file-sync": "^1.1.0",
     "path-exists": "^1.0.0",
     "path-is-absolute": "^1.0.0",
-    "source-map": "^0.4.0",
+    "source-map": "^0.5.0",
     "slash": "^1.0.0"
   },
   "bin": {