--- a/packages/babel/package.json
+++ b/packages/babel/package.json
@@ -69,7 +69,7 @@
     "resolve": "^1.1.6",
     "shebang-regex": "^1.0.0",
     "slash": "^1.0.0",
-    "source-map": "^0.4.0",
+    "source-map": "^0.5.0",
     "source-map-support": "^0.2.10",
     "to-fast-properties": "^1.0.0",
     "trim-right": "^1.0.0",