This PR should be merged simultaneously with https://github.com/babel/babel/pull/2424, so there aren't multiple versions of `source-map` installed. Notable changes include:

* Builds are run through [derequire](https://github.com/calvinmetcalf/derequire) and [bundle-collapser](https://github.com/substack/bundle-collapser). So they're smaller and safe to rebundle (if that's your thing).
* Minified files are compressed and mangled – that's something you have to specify in the CLI.
* Added a minified version of the node bundle. I ran some quick tests and using that instead of babel-core shaves 2 seconds off my builds. But, it does mean that the package then has another 860K to carry around. **Feel free to remove this if you want**
* Switched all the flags to their long-form so you know what they do. Also turned on bash debugging (`set -x`) as a poor-man's progress indicator.

| before | after | file |
| -----: | ----: | ---- |
| 2.1M   | 1.9M  | browser.js |
| 1.3M   | 871K  | browser.min.js |
|  12K   |  12K  | external-helpers.js |
| 8.7K   | 5.7K  | external-helpers.min.js |
| 2.0M   | 1.9M  | node.js |
|  n/a   | 860K  | node.min.js |
| 151K   | 135K  | polyfill.js |
| 103K   | 55K   | polyfill.min.js |