@zertosh 2015-09-24
Oh, the reason for doing this now is because of https://github.com/mozilla/source-map/pull/205 – they dropped their AMD wrapper thing so running bundle-collapser doesn't break it.